Schema Refactoring

+++ database/migrations/2021_07_02_152249_create_tours_table.php
@@ -18,7 +18,7 @@ class CreateToursTable extends Migration
             $table->string('title');
             $table->mediumText('description');
             $table->string('image');
-            $table->unsignedBigInteger('created_by');
+            $table->unsignedInteger('created_by');
             $table->string('location');
             $table->decimal('price')->default("0.00");
             $table->string('distance')->nullable();
