+++ database/migrations/2021_07_02_152504_create_highlights_table.php
@@ -20,7 +20,7 @@ class CreateHighlightsTable extends Migration
             $table->mediumText('video');
             $table->string('slug');
             $table->mediumText('description')->nullable();
-            $table->unsignedBigInteger('created_by');
+            $table->unsignedInteger('created_by');
             $table->foreign('created_by')->references('id')->on('users')
             ->onDelete('cascade')
             ->onUpdate('cascade');
